requirements.txt
autopep8==2.3.2
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
pycodestyle==2.14.0
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
python-dotenv==1.1.1
waitress==3.0.2
